make: delete compiling flag '-Wno-sign-compare' delete the '-Wno-sign-compare' of both CFLAGS and CXXFLAGS in file configure.ac, and work out the compiling error in comparison between signed and unsigned integer expression. I still use some type converion(convert 'int' into 'uint'), because changing the variabe type from its definition need too many code modifications.

+++ decoder/vaapidecoder_jpeg.cpp
@@ -307,8 +307,8 @@ Decode_Status VaapiDecoderJpeg::decodePictureStart()
 
     if (!m_hasContext
         || m_configBuffer.profile != profile
-        || (uint32_t)(m_configBuffer.width) != m_width
-        || (uint32_t)(m_configBuffer.height) != m_height) {
+        || m_configBuffer.width != m_width
+        || m_configBuffer.height != m_height) {
         m_configBuffer.surfaceNumber = 2;
         m_configBuffer.profile = profile;
         m_configBuffer.width = m_width;
